requirements.txt
# ============================
# UI & App Framework
# ============================
streamlit==1.56.0          # Core UI framework for PackPal, runs on HF Spaces
pandas==3.0.2              # Data manipulation for packing lists, schedules
numpy==2.4.4              # Numerical backbone for math + CV preprocessing
plotly==6.7.0              # Interactive charts for CLO scores & evaluation

# ============================
# CV & Weather 
# ============================
ultralytics==8.4.37        # YOLOv11/YOLO26 segmentation models
# --- HF Spaces Optimization: Install CPU versions only ---
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.11.0+cpu               # PyTorch deep learning framework
torchvision==0.26.0+cpu        # Image transforms & pretrained CV utilities
# ---------------------------------------------------------
opencv-python-headless==4.13.0.92  # Image processing (headless for HF Spaces)
requests==2.33.1           # Fetch APIs (e.g., weather data)
pillow==12.2.0             # Image manipulation (resize, crop, augment)
sentence-transformers==5.4.1 # Embeddings for slot detection & semantic similarity
scikit-image==0.26.0       # REQUIRED for LIME image explanations & preprocessing
transformers==5.5.4     # CLIP clothing analyser                 (--vision clip)
lightgbm==4.6.0 		# Gradient boosting framework for ML tasks.
segment-anything==1.0.0  # Meta's SAM model for image segmentation (YOLO alternative)

# ============================
# Optimization & Math
# ============================
ortools==9.15.6755         # Google OR-Tools, 3D bin packing solver
deap==1.4.3                # Genetic algorithms (NSGA-II for multi-objective optimization)
scipy==1.17.1              # Scientific computing, CLO thermal calculations
scikit-learn==1.8.0        # ML utilities (normalization, clustering, evaluation)
pygad==3.6.0          	   # NSGA-II genetic algorithm

# ============================
# Logic, Graph & NLU
# ============================
langchain==1.2.15           # LLM orchestration framework
langchain-groq==1.1.2      # Groq Cloud API integration
groq==0.37.1               # Explicitly pin the base Groq client
langchain-community==0.4.1 # Community-contributed LangChain modules
pydantic==2.13.1            # Schema enforcement for structured slot filling
neo4j==6.1.0              # Graph DB client for packing rules
supabase==2.25.1           # SQL + S3 storage client
boto3==1.42.89              # AWS SDK, required for Cloudflare R2 connection
python-dotenv==1.2.2       # Loads .env secrets safely
google-genai               # Gemini AI narrative advice (needs GEMINI_API_KEY env var)

# ============================
# XAI & Evaluation
# ============================
shap==0.51.0               # Explainable AI (global feature importance)
lime==0.2.0.1              # Local model explanations (image/text tabular)
matplotlib==3.10.8         # Static plots
seaborn==0.13.2            # Statistical visualization

# ============================
# Logic, Graph & NLU
# ============================
rich==15.0.0 		# Rich text formatting for Streamlit console/logs.
joblib==1.5.3 		# Parallelization + model persistence.
pyarrow==23.0.1 		# Arrow format for fast data interchange.

